Fix: unwrap function-call envelope en 3-expertos + score float en Delphi
- _parse: detectar wrapper {"type":"function","parameters":{...}} y extraer payload
- DimensionScore.score: int → float para aceptar promedios como 6.4

+++ docker-deploy/delphi/live_schema.py
@@ -17,7 +17,7 @@ DIMENSIONS = [
 
 class DimensionScore(BaseModel):
     dimension: str
-    score: int = Field(ge=1, le=10)
+    score: float = Field(ge=1, le=10)
     justificacion: str
